Attempted to make a first page
Hi guys, i started creating the first page of our website, but if you look at the code it says that i added the picture EuropePolitical to it, just so we had an idea of what it would look like, but when I open the page it comes up as a box with a question mark. Does anyone know why? Also does anyone know how to add a picture background of CSS as every time I have tried it hasn't worked. Thanks!